Redone parts of temp content transition and added compatibility for Portal Stories: Mel
Redone some parts of working with the new temp content folder system. It will properly save any preexisting temp content folders. However only if there is one, as the launcher will crash. I could have  made a system where it would iterate a number for the rename, but then there would have to be another max. Plus undoing it all would be ridiculous.

Added support for the launcher to properly load up Portal Stories: Mel. All that needs to be done client side is to change the game path, launcher takes over from there. Aperture Tag is also there, but commented out because I am not sure this will be finished in time for release.