Reworked how map support code is loaded
Made loaded map support code depending on the Portal 2 based game more seamless as well as adding the footing for support of other Portal 2 based games. g_iCurGameIndex will be used to indicate which Portal 2 based game is being run and the supplied constants for checks with specific games.

Also changed UTIL_PlayerByIndex to PlayerByIndex and added a for loop to remove logic_autosaves as they are useless with P2:MM.